refactor: extract history-readiness/hydration computation from app.py into core/history_readiness.py
Structural extraction — NO behavior change. Full suite 2268 passed/1 skipped,
identical to baseline; 77 characterization+architecture tests for this cluster
pass unchanged.

What & why: app.py was ~13.3k lines mixing orchestration with a large, cohesive
readiness/hydration *computation* layer. Moved that layer (520 lines) into a new
focused module so app.py keeps orchestration while the pure decision logic lives
on its own, next to its siblings core/history_roles.py (role SSOT) and
execution/readiness.py (contracts).

Moved verbatim (AST-verified identical bodies) into core/history_readiness.py:
- helpers: _count_bars_from_provider, _get_bars_from_provider, _bar_timestamp,
  _get_cached_quote, _status_for_role, _count_symbol_bars,
  _pick_atm_option_symbols_from_basket
- dataclasses: HistoryPolicyDecision, RuntimeHistoryResult, HistoryReadiness,
  SelectedOptionHistoryReadiness
- functions: build_symbol_hydration_status, _hydration_status_map,
  compute_history_readiness, resolve_symbol_history_role (ctx-adapter),
  compute_selected_option_history_readiness, resolve_history_policy

Compatibility: app.py re-imports every moved name, so all existing callers
(strategies/runner.py and ~71 tests importing from nifty_scalper_bot.core.app)
work unchanged. Two app.py-internal helpers used once each by the cluster
(ensure_bot_context_runtime_fields, _symbol_history_requirement) stay in app.py
and are imported lazily inside the two call sites to avoid a circular import;
the lazy path is runtime-exercised and works.

Test update: test_canonical_readiness_function_exists_and_is_pure now reads the
new module path (its intent — the canonical readiness function must be pure — is
unchanged; only the file location moved).

Verification: AST diff confirms exactly the 20 cluster definitions left app.py
with zero additions; the new module's function bodies are byte-identical to the
originals except the 2 deferred-import lines; no circular import; full suite green.
